+++ openseries/series.py
@@ -15,19 +15,28 @@ from pydantic import BaseModel, constr, Field, root_validator
 from re import compile
 from scipy.stats import kurtosis, norm, skew
 from stdnum import isin as isincode
-from typing import List, Literal, Tuple, TypeVar
+from typing import List, Tuple
 
 from openseries.datefixer import date_offset_foll, date_fix, holiday_calendar
 from openseries.exceptions import FromFixedRateDatesInputError
 from openseries.load_plotly import load_plotly_dict
+from openseries.types import (
+    Lit_quantile_interpolation,
+    Lit_bizday_frequencies,
+    Lit_pandas_resample_convention,
+    Lit_pandas_reindex_method,
+    Lit_nan_method,
+    Lit_line_plot_mode,
+    Lit_bar_plot_mode,
+    Lit_plotly_output,
+    TOpenTimeSeries,
+)
 from openseries.risk import (
     cvar_down,
     var_down,
     drawdown_series,
     drawdown_details,
 )
-
-TOpenTimeSeries = TypeVar("TOpenTimeSeries", bound="OpenTimeSeries")
 
 
 def check_if_none(item) -> bool:
@@ -1529,9 +1538,7 @@ class OpenTimeSeries(BaseModel):
     def var_down(
         self: TOpenTimeSeries,
         level: float = 0.95,
-        interpolation: Literal[
-            "linear", "lower", "higher", "midpoint", "nearest"
-        ] = "lower",
+        interpolation: Lit_quantile_interpolation = "lower",
     ) -> float:
         """Downside Value At Risk, "VaR". The equivalent of
         percentile.inc([...], 1-level) over returns in MS Excel \n
@@ -1542,8 +1549,7 @@ class OpenTimeSeries(BaseModel):
 
         level: float, default: 0.95
             The sought VaR level
-        interpolation: Literal["linear", "lower", "higher", "midpoint",
-        "nearest"], default: "lower"
+        interpolation: Lit_quantile_interpolation, default: "lower"
             type of interpolation in Pandas.DataFrame.quantile() function.
 
         Returns
@@ -1562,9 +1568,7 @@ class OpenTimeSeries(BaseModel):
         months_from_last: int | None = None,
         from_date: dt.date | None = None,
         to_date: dt.date | None = None,
-        interpolation: Literal[
-            "linear", "lower", "higher", "midpoint", "nearest"
-        ] = "lower",
+        interpolation: Lit_quantile_interpolation = "lower",
     ) -> float:
         """https://www.investopedia.com/terms/v/var.asp
         Downside Value At Risk, "VaR". The equivalent of
@@ -1582,8 +1586,7 @@ class OpenTimeSeries(BaseModel):
             Specific from date
         to_date : datetime.date, optional
             Specific to date
-        interpolation: Literal["linear", "lower", "higher", "midpoint",
-        "nearest"], default: "lower"
+        interpolation: Lit_quantile_interpolation, default: "lower"
             type of interpolation in Pandas.DataFrame.quantile() function.
 
         Returns
@@ -1603,9 +1606,7 @@ class OpenTimeSeries(BaseModel):
     def vol_from_var(
         self: TOpenTimeSeries,
         level: float = 0.95,
-        interpolation: Literal[
-            "linear", "lower", "higher", "midpoint", "nearest"
-        ] = "lower",
+        interpolation: Lit_quantile_interpolation = "lower",
     ) -> float:
         """
         Parameters
@@ -1613,8 +1614,7 @@ class OpenTimeSeries(BaseModel):
 
         level: float, default: 0.95
             The sought VaR level
-        interpolation: Literal["linear", "lower", "higher", "midpoint",
-        "nearest"], default: "lower"
+        interpolation: Lit_quantile_interpolation, default: "lower"
             type of interpolation in Pandas.DataFrame.quantile() function.
 
         Returns
@@ -1636,9 +1636,7 @@ class OpenTimeSeries(BaseModel):
         months_from_last: int | None = None,
         from_date: dt.date | None = None,
         to_date: dt.date | None = None,
-        interpolation: Literal[
-            "linear", "lower", "higher", "midpoint", "nearest"
-        ] = "lower",
+        interpolation: Lit_quantile_interpolation = "lower",
         drift_adjust: bool = False,
         periods_in_a_year_fixed: int | None = None,
     ) -> float:
@@ -1655,8 +1653,7 @@ class OpenTimeSeries(BaseModel):
             Specific from date
         to_date : datetime.date, optional
             Specific to date
-        interpolation: Literal["linear", "lower", "higher", "midpoint",
-        "nearest"], default: "lower"
+        interpolation: Lit_quantile_interpolation, default: "lower"
             type of interpolation in Pandas.DataFrame.quantile() function.
         drift_adjust: bool, default: False
             An adjustment to remove the bias implied by the average return
@@ -1711,9 +1708,7 @@ class OpenTimeSeries(BaseModel):
         months_from_last: int | None = None,
         from_date: dt.date | None = None,
         to_date: dt.date | None = None,
-        interpolation: Literal[
-            "linear", "lower", "higher", "midpoint", "nearest"
-        ] = "lower",
+        interpolation: Lit_quantile_interpolation = "lower",
         drift_adjust: bool = False,
         periods_in_a_year_fixed: int | None = None,
     ) -> float:
@@ -1737,8 +1732,7 @@ class OpenTimeSeries(BaseModel):
             Specific from date
         to_date : datetime.date, optional
             Specific to date
-        interpolation: Literal["linear", "lower", "higher", "midpoint",
-        "nearest"], default: "lower"
+        interpolation: Lit_quantile_interpolation, default: "lower"
             type of interpolation in Pandas.DataFrame.quantile() function.
         drift_adjust: bool, default: False
             An adjustment to remove the bias implied by the average return
@@ -1826,7 +1820,7 @@ class OpenTimeSeries(BaseModel):
         OpenTimeSeries
             An OpenTimeSeries object
         """
-
+        self.tsdf.columns: MultiIndex
         if not any(
             [
                 True if x == ValueType.RTRN else False
@@ -1894,11 +1888,9 @@ class OpenTimeSeries(BaseModel):
 
     def resample_to_business_period_ends(
         self: TOpenTimeSeries,
-        freq: Literal["BM", "BQ", "BA"] = "BM",
-        convention: Literal["start", "s", "end", "e"] = "end",
-        method: Literal[
-            None, "pad", "ffill", "backfill", "bfill", "nearest"
-        ] = "nearest",
+        freq: Lit_bizday_frequencies = "BM",
+        convention: Lit_pandas_resample_convention = "end",
+        method: Lit_pandas_reindex_method = "nearest",
     ) -> TOpenTimeSeries:
         """Resamples timeseries frequency to the business calendar
         month end dates of each period while leaving any stubs
@@ -1906,12 +1898,11 @@ class OpenTimeSeries(BaseModel):
 
         Parameters
         ----------
-        freq: Literal["BM", "BQ", "BA"], default BM
+        freq: Lit_bizday_frequencies, default BM
             The date offset string that sets the resampled frequency
-        convention: Literal["start", "s", "end", "e"], default; end
+        convention: Lit_pandas_resample_convention, default; end
             Controls whether to use the start or end of `rule`.
-        method: Literal[None, "pad", "ffill", "backfill", "bfill",
-        "nearest"], default: nearest
+        method: Lit_pandas_reindex_method, default: nearest
             Controls the method used to align values across columns
 
         Returns
@@ -2131,9 +2122,7 @@ class OpenTimeSeries(BaseModel):
         self: TOpenTimeSeries,
         level: float = 0.95,
         observations: int = 252,
-        interpolation: Literal[
-            "linear", "lower", "higher", "midpoint", "nearest"
-        ] = "lower",
+        interpolation: Lit_quantile_interpolation = "lower",
     ) -> DataFrame:
         """
         Parameters
@@ -2142,8 +2131,7 @@ class OpenTimeSeries(BaseModel):
             The sought Value At Risk level
         observations: int, default: 252
             Number of observations in the overlapping window.
-        interpolation: Literal["linear", "lower", "higher", "midpoint",
-        "nearest"], default: "lower"
+        interpolation: Lit_quantile_interpolation, default: "lower"
             Type of interpolation in Pandas.DataFrame.quantile() function.
 
         Returns
@@ -2161,13 +2149,13 @@ class OpenTimeSeries(BaseModel):
         return vardf
 
     def value_nan_handle(
-        self: TOpenTimeSeries, method: Literal["fill", "drop"] = "fill"
+        self: TOpenTimeSeries, method: Lit_nan_method = "fill"
     ) -> TOpenTimeSeries:
         """Handling of missing values in a valueseries
 
         Parameters
         ----------
-        method: Literal["fill", "drop"], default: "fill"
+        method: Lit_nan_method, default: "fill"
             Method used to handle NaN. Either fill with last known or drop
 
         Returns
@@ -2187,13 +2175,13 @@ class OpenTimeSeries(BaseModel):
         return self
 
     def return_nan_handle(
-        self: TOpenTimeSeries, method: Literal["fill", "drop"] = "fill"
+        self: TOpenTimeSeries, method: Lit_nan_method = "fill"
     ) -> TOpenTimeSeries:
         """Handling of missing values in a returnseries
 
         Parameters
         ----------
-        method: Literal["fill", "drop"], default: "fill"
+        method: Lit_nan_method, default: "fill"
             Method used to handle NaN. Either fill with zero or drop
 
         Returns
@@ -2230,7 +2218,7 @@ class OpenTimeSeries(BaseModel):
         OpenTimeSeries
             An OpenTimeSeries object
         """
-
+        self.tsdf.columns: MultiIndex
         if any(
             [
                 True if x == ValueType.RTRN else False
@@ -2307,19 +2295,19 @@ class OpenTimeSeries(BaseModel):
 
     def plot_series(
         self: TOpenTimeSeries,
-        mode: Literal["lines", "markers", "lines+markers"] = "lines",
+        mode: Lit_line_plot_mode = "lines",
         tick_fmt: str | None = None,
         directory: str | None = None,
         auto_open: bool = True,
         add_logo: bool = True,
         show_last: bool = False,
-        output_type: Literal["file", "div"] = "file",
+        output_type: Lit_plotly_output = "file",
     ) -> Tuple[Figure, str]:
         """Creates a Plotly Figure
 
         Parameters
         ----------
-        mode: Literal["lines", "markers", "lines+markers"], default: "lines"
+        mode: Lit_line_plot_mode, default: "lines"
             The type of scatter to use
         tick_fmt: str, optional
             None, '%', '.1%' depending on number of decimals to show
@@ -2331,8 +2319,8 @@ class OpenTimeSeries(BaseModel):
             If True a Captor logo is added to the plot
         show_last: bool, default: False
             If True the last data point is highlighted as red dot with a label
-        output_type: str, default: "file"
-            file or div
+        output_type: Lit_plotly_output, default: "file"
+            Determines output type
 
         Returns
         -------
@@ -2398,16 +2386,19 @@ class OpenTimeSeries(BaseModel):
 
     def plot_bars(
         self: TOpenTimeSeries,
+        mode: Lit_bar_plot_mode = "group",
         tick_fmt: str | None = None,
         directory: str | None = None,
         auto_open: bool = True,
         add_logo: bool = True,
-        output_type: Literal["file", "div"] = "file",
+        output_type: Lit_plotly_output = "file",
     ) -> Tuple[Figure, str]:
         """Creates a Plotly Bar Figure
 
         Parameters
         ----------
+        mode: Lit_bar_plot_mode, default: "group"
+            The type of bar to use
         tick_fmt: str, optional
             None, '%', '.1%' depending on number of decimals to show
         directory: str, optional
@@ -2416,8 +2407,8 @@ class OpenTimeSeries(BaseModel):
             Determines whether to open a browser window with the plot
         add_logo: bool, default: True
             If True a Captor logo is added to the plot
-        output_type: str, default: "file"
-            file or div
+        output_type: Lit_plotly_output, default: "file"
+            Determines output type
 
         Returns
         -------
@@ -2437,7 +2428,7 @@ class OpenTimeSeries(BaseModel):
             hovertemplate="%{y}<br>%{x|%Y-%m-%d}",
             name=self.label,
         )
-        figure.update_layout(yaxis=dict(tickformat=tick_fmt))
+        figure.update_layout(barmode=mode, yaxis=dict(tickformat=tick_fmt))
 
         if add_logo:
             figure.add_layout_image(logo)